pyproject.toml
[tool.poetry]
package-mode = false
name = "poster-tg"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "poster"}]

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
pytelegrambotapi = "^4.22.1"
ruff = "^0.6.5"
mypy = "^1.11.2"
pydantic = "^2.9.2"
pytest = "^8.3.3"

[tool.pytest.ini_options]
pythonpath = [".",]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
stubPath = "/stubs"
venv= ".venv"
venvPath = "."
reportUnusedVariable = false
reportUnusedImport = false
severityLevel = "none"
ignore= [
  "stubs/",
  "alembic/",
  "env_.py",
  "easy_p2p/__main__.py",
]

[tool.isort]
import_heading_stdlib= "Standard Library"
import_heading_thirdparty= "Third Party Stuff"
import_heading_firstparty= "My Stuff"

multi_line_output= 5
line_length= 60
combine_as_imports= "True"
include_trailing_comma= "True"
force_grid_wrap= 2

[tool.ruff]
# https://docs.astral.sh/ruff/rules/
lint.select = [
  "F", # flake8
  "E", # pycodestyle errors
  "W", # pycodestyle warnings
  "B", # flake8-bugbear
  "N", # Check your code against PEP 8 naming conventions.
  "C", # McCabe complexity.
  "T", # Print found todos.
  # "D", # Docstring checks.
  "UP", # upgrade syntax for newer versions of the language.
  "ANN", # Check for missing type annotations 
  "S", # Check for security issues
  "DTZ", # Check for datetime timezone issues
]
lint.ignore = [
  "ANN002", # Missing type annotation for *args
  "ANN003", # Missing type annotation for **kwargs
  "ANN101", # Missing type annotation for self in method
  "ANN204", # Missing type annotation for special methods like __init__ etc.
  "ANN102", # Missing type annotation for cls in classmethod
  "S101", # Use of assert detected
  # "E712", # comparison-with-callable use "is" instead of "=="
  # "E711", # comparison-with-callable use "is" instead of "=="
  # "E501", # line too long
]
# 3. Avoid trying to fix flake8-bugbear (`B`) violations.
# lint.unfixable = ["B"]
exclude = [
  'stubs/',
  'trash/',
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
  "S101", # disable security issues for tests
  "ANN201",
]
"core/mock/*" = [
  "S101", # disable security issues for tests
]

[tool.ruff.format]
# https://docs.astral.sh/ruff/settings/#format_skip-magic-trailing-comma
skip-magic-trailing-comma = false

[tool.mypy]
# ignore_missing_imports = true
# no_implicit_optional = false
mypy_path = "$MYPY_CONFIG_FILE_DIR/stubs"
exclude = ['stubs/', 'trash/']
